+++ modules/account/mocks/mockUserAvatarImage.ts
