+++ modules/post/components/detail/planner/Planner.tsx
@@ -3,12 +3,12 @@ import { View } from 'react-native';
 
 import { ScreenTitle } from '@shared/components/font/Font';
 import ScreenLayout from '@shared/components/layout/ScreenLayout';
+import DividerWrapper from '@shared/components/layout/DividerWrapper';
 import Map from './map/Map';
 import VenueList from './venueList/VenueList';
 import AddVenue from './addVenue/AddVenue';
 import { Iplanner } from './type';
 import ContentHeader from '../ContentHeader';
-import DividerWrapper from '../../../../shared/components/layout/DividerWrapper';
 import useStyles from './style';
 
 const TITLE = '모임 장소';
